fix: resolved code reviewer comments
